+++ spec/yandex_cloud/translate_spec.rb
@@ -14,10 +14,10 @@ RSpec.describe YandexCloud::Translate do
   end
 
   describe 'methods' do
+    let!(:bad_translate_service) { YandexCloud::Translate.new(iam_token: iam_token, folder_id: '12345') }
+
     context '.languages' do
       context 'for invalid request' do
-        let!(:bad_translate_service) { YandexCloud::Translate.new(iam_token: iam_token, folder_id: '12345') }
-
         it 'returns response with error' do
           response = bad_translate_service.languages
 
@@ -30,6 +30,30 @@ RSpec.describe YandexCloud::Translate do
           response = translate_service.languages
 
           expect(response['languages'].is_a?(Array)).to eq true
+        end
+      end
+    end
+
+    context '.detect' do
+      context 'for invalid request' do
+        it 'returns response with error' do
+          response = bad_translate_service.detect(text: 'Hola')
+
+          expect(response['error_message']).to_not eq nil
+        end
+      end
+
+      context 'for valid request' do
+        it 'returns name of language, without hint' do
+          response = translate_service.detect(text: 'Hola')
+
+          expect(response['language']).to eq 'es'
+        end
+
+        it 'returns name of language, with hint' do
+          response = translate_service.detect(text: 'Hola', hint: 'es')
+
+          expect(response['language']).to eq 'es'
         end
       end
     end